form update

+++ database/migrations/2021_02_19_140204_create_forms_table.php
@@ -98,7 +98,7 @@ class CreateFormsTable extends Migration
 
             $table->date('dateOnsetOfIllness')->nullable();
             $table->text('SAS')->nullable();
-            $table->mediumInteger('SASFeverDeg')->nullable();
+            $table->double('SASFeverDeg')->nullable();
             $table->text('SASOtherRemarks')->nullable();
             $table->text('COMO');
             $table->text('COMOOtherRemarks')->nullable();
